Expose ACK rate for Semtech UDP backend per Gateway ID and context in Error Event
Details:
- Also expose Extra information (context) to Error Event
If the Gateway Bridge is installed on the server and
especially using Docker, there can be cases where the uplink path (GW >
CGB) works, but the downlink path (CGB > GW) fails (UDP packets never
reach the GW). By exposing the ACK rate per Gateway ID, it is possible
to add per gateway monitoring.

As the exposed ack-rate represents the last reported ack-rate by the GW,
it is also important to monitor the ack-rate count. If this value does
not increment, it could mean that the GW is no longer reporting the
ack-rate and thus the exposed ack-rate is stale.

Signed-off-by: SAGAR PATEL <sagar.a.patel@slscorp.com>